ftp-receiver/requirements.txt
loguru==0.7.2
pika==1.3.2
python-decouple==3.8
